The Warrior Princess

+++ util.go
@@ -4,10 +4,10 @@ import "fmt"
 
 var unit = []string{"", "kB", "MB", "GB", "TB", "PB"}
 
-func FormatSize(n uint64) string {
+func FormatSize(n float64) string {
 	i := 0
-	for ; n > 1000; i++ {
-		n /= 1000
+	for ; n > 1000.0; i++ {
+		n /= 1000.0
 	}
-	return fmt.Sprintf("%d %s", n, unit[i])
+	return fmt.Sprintf("%5.1f %s", n, unit[i])
 }
